Moving WebhookWriter to separate file
